第二天

+++ src/Algorithms/Node.java
@@ -1,13 +1,14 @@
 package Algorithms;
 
-import station_database.Station;
+import Station_Data.Station;
 
+//注：该Node类用于存储open中站点的g，f值。
 public class Node {
     public Station station;
-    public int g;  // 起点到当前点的代价
-    public int f;  // f = g + h
+    public double g;  // 起点到当前点的代价
+    public double f;  // f = g + h
 
-    public Node(Station station, int g, int f) {
+    public Node(Station station, double g, double f) {
         this.station = station;
         this.g = g;
         this.f = f;
